Code review
Licensing and copying - was automatically generated when I committed from Visual Studio, obviously I’m not releasing this to world. The license has now been included in the code
C++11 – Will need to check this, if not some changes.
size_t for size parameters, the existing SAS code base uses int for data lengths so I’m avoiding a conversion here (of course negative is garbage) Should I make it more explicit or just leave it? (Will throw random unsigned conversion warnings in their code)
Hash functions getting seed from PRNG, I have tested 1 through n as seeds and the false positive rate has degraded by about 5% consistently, as does similar strategies of k*LARGE_CONSTANT – since we’re only generating 10 numbers here and using a new prng each bloom filter – the probability of getting two the same should be exactly zero (Need to worry about this? Maybe hard-code an array if paranoid)
Hash function occupancy – the hash functions define a 2d array T bytes_[N][byte_range] which they u

+++ main.cpp
@@ -3,7 +3,7 @@
 #include <random>
 #include <string>
 #include <vector>
-#include "marker_cache.h"
+#include "markercache.h"
 
 using namespace std;
 
@@ -18,7 +18,7 @@ vector<char*> generate_test_data(size_t num_elems, size_t width) {
         "!@#$%^&*()"
         "`~-_=+[{]}\\|;:'\",<.>/? ");
     default_random_engine rng(random_device{}());
-    uniform_int_distribution<int> distribution(0, chars.size() - 1);
+    uniform_int_distribution<size_t> distribution(0, chars.size() - 1);
     for (auto i = 0; i < num_elems; ++i) {
         auto s = new char[width];
         for (size_t i = 0; i < width; ++i) s[i] = chars[distribution(rng)];
